+++ app/Gateways/UsersGateway.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace App\Gateways;
+
+use App\Interfaces\UsersInterface;
+
+class UsersGateway extends BaseGateway
+{
+    public function __construct(UsersInterface $interface)
+    {
+        $this->setInterface($interface);
+        ;
+    }
+}