+++ app/Http/Controllers/ApiController.php
@@ -7,6 +7,11 @@ class ApiController extends Controller
 {
     private $gateway;
 
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
     protected function setGateway($gateway)
     {
         $this->gateway = $gateway;