+++ app/Interfaces/UsersInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+
+namespace App\Interfaces;
+
+interface UsersInterface extends BaseInterface
+{
+}