+++ database/seeders/DatabaseSeeder.php
@@ -2,6 +2,8 @@
 
 namespace Database\Seeders;
 
+use App\Repositories\RolesRepository;
+use App\Repositories\UsersRepository;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -13,6 +15,9 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        // \App\Models\User::factory(10)->create();
+        // Admin role with id 1
+        RolesRepository::factory()->create();
+        // Admin user
+        UsersRepository::factory()->create();
     }
 }